backend/requirements.txt
aiohttp==1.1.6
aiohttp-cors==0.5.0
aiohttp-jinja2==0.12.0
appnope==0.1.0
async-timeout==1.1.0
biopython==1.68
bleach==1.5.0
cffi==1.9.1
chardet==2.3.0
cycler==0.10.0
decorator==4.0.11
dill==0.2.5
empath==0.89
entrypoints==0.2.2
future==0.16.0
gprofiler-official==0.3.4
html5lib==0.999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
Jinja2==2.9.6
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.4
multidict==2.1.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
numpy==1.11.2
oauthlib==2.0.2
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
protobuf==3.1.0
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.2.0
PyInstaller==3.2.1
pyparsing==2.1.4
python-dateutil==2.6.0
python-twitter==3.2.1
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.5
requests-oauthlib==0.8.0
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==0.12.0rc1
terminado==0.6
testpath==0.3
torch==0.1
torchvision==0.1.7
tornado==4.4.2
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
widgetsnbextension==2.0.0
yarl==0.8.1
